backend/requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Spond Integration
spond==1.1.1
spond-classes==0.17.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==44.0.0

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Background Tasks & Scheduling
apscheduler==3.10.4

# HTTP Client
httpx==0.28.1
aiofiles==24.1.0

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.9.0.post0
